fix: :bug: fix feedback rating type to float64 to match database

+++ src/models/feedback.entity.go
@@ -4,6 +4,6 @@ package models
 // User represents a user entity
 type Feedback struct {
 	FeedbackID string  `json:"feedbackID" bson:"feedbackID"`
-	Rating     float32 `json:"rating" bson:"rating"`
+	Rating     float64 `json:"rating" bson:"rating"`
 	Content    string  `json:"content" bson:"content"`
 }
